backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# LangGraph + LangChain ecosystem
langgraph>=0.4.0
langchain-core>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# AWS
boto3>=1.35.0

# Vector DB + Embeddings
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# PDF processing (vectordb creation pipeline)
PyMuPDF>=1.25.0

# Optional: only needed for vectordb_util summarization pipeline
# transformers
# torch
